refactor: Re-use client in consumer
