+++ hook-consumer/src/lib.rs
@@ -1,2 +1,3 @@
 pub mod config;
 pub mod consumer;
+pub mod error;
